Renamed scss files to  styles.scss
