+++ src/app/screens/App/screens/Home/components/FilterSelect/styles.scss
@@ -1,4 +1,4 @@
-@import '../../../../../app';
+@import '../../../../../../app';
 
 .filter-select {
   @extend %input-styles;
