+++ src/app/screens/App/screens/Home/components/SearchBox/styles.scss
@@ -1,4 +1,4 @@
-@import '../../../../../app';
+@import '../../../../../../app';
 
 .search-container {
   display: flex;
