+++ src/app/screens/App/screens/Home/styles.scss
